Rename wordstore* -> word_store*.
