+++ searchlib/src/tests/memoryindex/datastore/CMakeLists.txt
@@ -6,10 +6,10 @@ vespa_add_executable(searchlib_feature_store_test_app TEST
     searchlib
 )
 vespa_add_test(NAME searchlib_feature_store_test_app COMMAND searchlib_feature_store_test_app)
-vespa_add_executable(searchlib_wordstore_test_app TEST
+vespa_add_executable(searchlib_word_store_test_app TEST
     SOURCES
-    wordstore_test.cpp
+    word_store_test.cpp
     DEPENDS
     searchlib
 )
-vespa_add_test(NAME searchlib_wordstore_test_app COMMAND searchlib_wordstore_test_app)
+vespa_add_test(NAME searchlib_word_store_test_app COMMAND searchlib_word_store_test_app)
