+++ searchlib/src/vespa/searchlib/memoryindex/word_store.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "wordstore.h"
+#include "word_store.h"
 #include <vespa/searchlib/datastore/datastore.hpp>
 
 namespace search::memoryindex {
