+++ searchlib/src/vespa/searchlib/memoryindex/word_store.h
